+++ Classes/Controller/Example/PagesController.php
@@ -39,6 +39,13 @@ class PagesController extends AbstractBackendController
         $this->tableConfiguration['pages']['columns']['description']['defaultPosition'] = 4;
     }
 
+    protected function getDefaultFilters(): array
+    {
+        return [
+            'doktype' => ['value' => '1', 'expr' => 'eq'],
+        ];
+    }
+
     public function modifyQueryBuilder(): void
     {
         $this->queryBuilder->addSelectLiteral('CASE WHEN t1.l10n_parent != 0 THEN t1.l10n_parent ELSE t1.uid END AS ' . $this->queryBuilder->quoteIdentifier('sys_language_ordering'));